requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
scikit-learn==1.7.1
numpy==2.3.2
openai==1.99.9
google-genai==1.35.0
python-dotenv==1.1.1
